move shared ws types to wsShared
